refactor(backend): refactor and clearly organize the code by domain and processing stage
Clearly separate the employees, jobsites, workblocks modules.
Clearly separate the routing, controller, service, repository levels of the "call chain".
