+++ backend/src/controllers/jobsiteController.js
@@ -1,0 +1,6 @@
+import { addJobsite, findJobsites, getDefaultJobsiteProperties, } from '../services/jobsiteService.js'
+
+export const getJobsites = async (req, res) => {
+  const result = await findJobsites(req.params.query);
+  res.json(result);
+}
