+++ backend/src/repositories/employeeRepository.js
@@ -1,0 +1,3 @@
+import addEmployeeRecord from './addEmployeeRecord.js';
+
+export { addEmployeeRecord };
