+++ backend/src/repositories/projectRepository.js
@@ -1,0 +1,5 @@
+import { findProjectRecords } from "./findProjectRecords.js";
+import { addProjectRecord } from "./addProjectRecord.js";
+import { getDefaultProjectPropertiesRecord } from "./getDefaultProjectPropertiesRecord.js";
+
+export { findProjectRecords, addProjectRecord, getDefaultProjectPropertiesRecord };
