+++ backend/src/routes/workPeriodRouter.js
@@ -1,7 +1,7 @@
 import express from 'express';
-import getWorkPeriods from '../controllers/workPeriodController.js';
+import { getWorkPeriods } from '../controllers/workPeriodController.js';
 
 const workPeriodRouter = express.Router();
 workPeriodRouter.get('/', getWorkPeriods);
 
-export default workPeriodRouter;
+export { workPeriodRouter };
