+++ backend/src/services/employeeService.js
@@ -1,0 +1,10 @@
+import { addEmployeeRecord } from '../repositories/addEmployeeRecord.js';
+
+export function addEmployee(
+  employee_name = null,
+  employee_nickname = null,
+  title = null,
+  email = null,
+) {
+  return addEmployeeRecord();
+}
